2011-06-13  Sergio Martin <sergio.martin@artica.es>
	* include/functions_graph.php
	include/graphs/functions_pchart.php
	include/graphs/functions_gd.php
	include/graphs/fgraph.php
	include/graphs/functions_utils.php
	include/functions_reporting.php
	include/functions_visual_map.php
	include/ajax/visual_console_builder.ajax.php
	include/config_process.php
	include/include_graph_dependencies.php
	operation/agentes/stat_win.php
	operation/reporting/reporting_xml.php: Fixed missed parameters
	in sparse graph functions.
	Fixed graph includes and bugs



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@4432 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/include/graphs/functions_utils.php
@@ -10,15 +10,15 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-function serialize_in_temp($array = array(), $serial_id = null) {
+function serialize_in_temp($array = array(), $serial_id = null, $ttl = 1) {
 	$json = json_encode($array);
 	
 	if ($serial_id === null) {
 		$serial_id = uniqid();
 	}
-	
-	$file_path = sys_get_temp_dir()."/pandora_serialize_".$serial_id;
-		
+
+	$file_path = sys_get_temp_dir()."/pandora_serialize_".$serial_id."__1__".$ttl;
+
 	if (file_put_contents($file_path, $json) === false) {
 		return false;
 	}
@@ -26,12 +26,21 @@ function serialize_in_temp($array = array(), $serial_id = null) {
 	return $serial_id;
 }
 
-function unserialize_in_temp($serial_id = null, $delete = true) {
+function unserialize_in_temp($serial_id = null, $delete = true, $ttl = 1) {
 	if ($serial_id === null) {
 		return false;
 	}
 	
-	$file_path = sys_get_temp_dir()."/pandora_serialize_".$serial_id;
+	$volume = -1;
+	
+	for($i = 1 ; $i <= $ttl ; $i++) {
+		$file_path = sys_get_temp_dir()."/pandora_serialize_".$serial_id."__".$i."__".$ttl;
+		
+		if(file_exists($file_path)) {
+			$volume = $i;
+			break;
+		}
+	}
 
 	$content = file_get_contents($file_path);
 
@@ -42,7 +51,13 @@ function unserialize_in_temp($serial_id = null, $delete = true) {
 	$array = json_decode($content, true);
 	
 	if ($delete) {
-		unlink($file_path);
+		if($volume == $ttl) {
+			unlink($file_path);
+		}
+		else {
+			$next_volume = $volume++;
+			rename($file_path, sys_get_temp_dir()."/pandora_serialize_".$serial_id."__".$next_volume."__".$ttl);
+		}
 	}
 
 	return $array;
